requirements.txt
mcp==1.9.0
requests==2.32.3
mcp[cli]==1.9.0
mysql-connector-python==9.3.0